split hooks and refactored component packages
