+++ app/components/ERC20/TokenBalance/index.tsx
@@ -1,4 +1,4 @@
-import { useToken } from "../../../hooks/contract";
+import { useToken } from "../../../hooks/erc20";
 import { Text, Heading, Card } from "@raidguild/design-system";
 import { ethers } from "ethers";
