+++ app/components/ERC20/index.ts
@@ -1,0 +1,6 @@
+import Approve from "./Approve";
+import MintToken from "./MintToken";
+import TokenBalance from "./TokenBalance";
+import TransferToken from "./TransferToken";
+
+export { Approve, MintToken, TokenBalance, TransferToken };
