+++ app/components/ERC721/MintNFT/index.tsx
@@ -1,4 +1,4 @@
-import { useNFTs } from "../../../hooks/contract";
+import { useNFTs } from "../../../hooks/erc721";
 import {
   Text,
   Heading,
