+++ app/components/Staking/WithdrawStake/index.tsx
@@ -1,4 +1,4 @@
-import { useStaking } from "../../../hooks/contract";
+import { useStaking } from "../../../hooks/staking";
 import {
   Text,
   Heading,
