+++ app/components/Staking/index.ts
@@ -1,0 +1,7 @@
+import AllStakes from "./AllStakes";
+import ClaimRewards from "./ClaimRewards";
+import DepositStake from "./DepositStake";
+import StakeOverview from "./StakeOverview";
+import WithdrawStake from "./WithdrawStake";
+
+export { AllStakes, ClaimRewards, DepositStake, StakeOverview, WithdrawStake };
